+++ rust/benches/models/mod.rs
@@ -1,0 +1,5 @@
+#![allow(non_snake_case)]
+
+pub mod parse_models;
+pub mod serialize_models;
+mod utils;
